requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.1.1
executing==2.1.0
Faker==37.1.0
greenlet==3.2.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
ipython==8.31.0
jedi==0.19.2
matplotlib-inline==0.1.7
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
playwright==1.51.0
pluggy==1.5.0
prompt_toolkit==3.0.49
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==12.1.1
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.3
selenium==4.15.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
